requirements.txt
python-telegram-bot==20.7
openai==1.33.0
requests
duckduckgo-search==4.4